+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and as of version 1.0.0, follows semantic versioning.
 
 ## [Unreleased]
+## Changed
+  * Use Total Clusters instead of Total Reads in single lane views
 
 ## [200921-1501] - 2020-09-21
 ## Changed